Simplified Queries

+++ src/main/java/week2/exercise/domain/DailySummary.java
@@ -9,13 +9,13 @@ public class DailySummary {
     public final String date;
     public final  double maxPrice;
     public final  double minPrice;
-    public final  int sumVolume;
+    public final  long sumVolume;
     public final  double closingPrice;
 
     @JsonCreator
     public DailySummary(@JsonProperty("symbol") String symbol, @JsonProperty("date") String date,
                         @JsonProperty("maxPrice") double maxPrice, @JsonProperty("minPrice") double minPrice,
-                        @JsonProperty("sumVolume") int sumVolume, @JsonProperty("closingPrice") double closingPrice){
+                        @JsonProperty("sumVolume") long sumVolume, @JsonProperty("closingPrice") double closingPrice){
         this.symbol = symbol;
         this.date = date;
         this.maxPrice = maxPrice;
